refactor: LineService find 수정
