+++ src/main/java/wooteco/subway/dto/response/LineResponse.java
@@ -1,14 +1,19 @@
 package wooteco.subway.dto.response;
 
+import java.util.List;
+
 public class LineResponse {
     private final Long id;
     private final String name;
     private final String color;
+    private final List<StationResponse> stations;
 
-    public LineResponse(Long id, String name, String color) {
+    public LineResponse(Long id, String name, String color,
+        List<StationResponse> stations) {
         this.id = id;
         this.name = name;
         this.color = color;
+        this.stations = stations;
     }
 
     public Long getId() {
@@ -21,5 +26,9 @@ public class LineResponse {
 
     public String getColor() {
         return color;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
     }
 }
