--- src/main/java/wooteco/subway/dto/response/LineResponse2.java
@@ -1,34 +1,0 @@
-package wooteco.subway.dto.response;
-
-import java.util.List;
-
-public class LineResponse2 {
-    private final Long id;
-    private final String name;
-    private final String color;
-    private final List<StationResponse> stations;
-
-    public LineResponse2(Long id, String name, String color,
-        List<StationResponse> stations) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.stations = stations;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public List<StationResponse> getStations() {
-        return stations;
-    }
-}
